requirements.txt
matplotlib
numpy 
zmq
pyserial